+++ action/protocol/staking/ethabi/stake_bucketscount.go
@@ -1,4 +1,4 @@
-package staking
+package ethabi
 
 import (
 	"encoding/hex"
