+++ action/protocol/staking/ethabi/stake_candidatebyaddress.go
@@ -1,25 +1,27 @@
-package staking
+package ethabi
 
 import (
 	"encoding/hex"
 	"strings"
 
 	"github.com/ethereum/go-ethereum/accounts/abi"
+	"github.com/ethereum/go-ethereum/common"
+	"github.com/iotexproject/iotex-address/address"
 	"github.com/iotexproject/iotex-proto/golang/iotexapi"
 	"github.com/iotexproject/iotex-proto/golang/iotextypes"
 	"google.golang.org/protobuf/proto"
 )
 
-const _candidateByNameInterfaceABI = `[
+const _candidateByAddressInterfaceABI = `[
 	{
 		"inputs": [
 			{
-				"internalType": "string",
-				"name": "candName",
-				"type": "string"
+				"internalType": "address",
+				"name": "ownerAddress",
+				"type": "address"
 			}
 		],
-		"name": "candidateByName",
+		"name": "candidateByAddress",
 		"outputs": [
 			{
 				"components": [
@@ -69,47 +71,51 @@ const _candidateByNameInterfaceABI = `[
 	}
 ]`
 
-var _candidateByNameMethod abi.Method
+var _candidateByAddressMethod abi.Method
 
 func init() {
-	_interface, err := abi.JSON(strings.NewReader(_candidateByNameInterfaceABI))
+	_interface, err := abi.JSON(strings.NewReader(_candidateByAddressInterfaceABI))
 	if err != nil {
 		panic(err)
 	}
 	var ok bool
-	_candidateByNameMethod, ok = _interface.Methods["candidateByName"]
+	_candidateByAddressMethod, ok = _interface.Methods["candidateByAddress"]
 	if !ok {
 		panic("fail to load the method")
 	}
 }
 
-// CandidateByNameStateContext context for CandidateByName
-type CandidateByNameStateContext struct {
+// CandidateByAddressStateContext context for candidateByAddress
+type CandidateByAddressStateContext struct {
 	*baseStateContext
 }
 
-func newCandidateByNameStateContext(data []byte) (*CandidateByNameStateContext, error) {
+func newCandidateByAddressStateContext(data []byte) (*CandidateByAddressStateContext, error) {
 	paramsMap := map[string]interface{}{}
 	ok := false
-	if err := _candidateByNameMethod.Inputs.UnpackIntoMap(paramsMap, data); err != nil {
+	if err := _candidateByAddressMethod.Inputs.UnpackIntoMap(paramsMap, data); err != nil {
 		return nil, err
 	}
-	var candName string
-	if candName, ok = paramsMap["candName"].(string); !ok {
+	var ownerAddress common.Address
+	if ownerAddress, ok = paramsMap["ownerAddress"].(common.Address); !ok {
 		return nil, errDecodeFailure
+	}
+	owner, err := address.FromBytes(ownerAddress[:])
+	if err != nil {
+		return nil, err
 	}
 
 	method := &iotexapi.ReadStakingDataMethod{
-		Method: iotexapi.ReadStakingDataMethod_CANDIDATE_BY_NAME,
+		Method: iotexapi.ReadStakingDataMethod_CANDIDATE_BY_ADDRESS,
 	}
 	methodBytes, err := proto.Marshal(method)
 	if err != nil {
 		return nil, err
 	}
 	arguments := &iotexapi.ReadStakingDataRequest{
-		Request: &iotexapi.ReadStakingDataRequest_CandidateByName_{
-			CandidateByName: &iotexapi.ReadStakingDataRequest_CandidateByName{
-				CandName: candName,
+		Request: &iotexapi.ReadStakingDataRequest_CandidateByAddress_{
+			CandidateByAddress: &iotexapi.ReadStakingDataRequest_CandidateByAddress{
+				OwnerAddr: owner.String(),
 			},
 		},
 	}
@@ -117,7 +123,7 @@ func newCandidateByNameStateContext(data []byte) (*CandidateByNameStateContext, 
 	if err != nil {
 		return nil, err
 	}
-	return &CandidateByNameStateContext{
+	return &CandidateByAddressStateContext{
 		&baseStateContext{
 			&Parameters{
 				MethodName: methodBytes,
@@ -128,7 +134,7 @@ func newCandidateByNameStateContext(data []byte) (*CandidateByNameStateContext, 
 }
 
 // EncodeToEth encode proto to eth
-func (r *CandidateByNameStateContext) EncodeToEth(resp *iotexapi.ReadStateResponse) (string, error) {
+func (r *CandidateByAddressStateContext) EncodeToEth(resp *iotexapi.ReadStateResponse) (string, error) {
 	var result iotextypes.CandidateV2
 	if err := proto.Unmarshal(resp.Data, &result); err != nil {
 		return "", err
@@ -139,7 +145,7 @@ func (r *CandidateByNameStateContext) EncodeToEth(resp *iotexapi.ReadStateRespon
 		return "", err
 	}
 
-	data, err := _candidateByNameMethod.Outputs.Pack(cand)
+	data, err := _candidateByAddressMethod.Outputs.Pack(cand)
 	if err != nil {
 		return "", nil
 	}
