+++ action/protocol/staking/ethabi/stake_candidatebyname.go
@@ -1,4 +1,4 @@
-package staking
+package ethabi
 
 import (
 	"encoding/hex"
@@ -10,21 +10,16 @@ import (
 	"google.golang.org/protobuf/proto"
 )
 
-const _candidatesInterfaceABI = `[
+const _candidateByNameInterfaceABI = `[
 	{
 		"inputs": [
 			{
-				"internalType": "uint32",
-				"name": "offset",
-				"type": "uint32"
-			},
-			{
-				"internalType": "uint32",
-				"name": "limit",
-				"type": "uint32"
+				"internalType": "string",
+				"name": "candName",
+				"type": "string"
 			}
 		],
-		"name": "candidates",
+		"name": "candidateByName",
 		"outputs": [
 			{
 				"components": [
@@ -64,9 +59,9 @@ const _candidatesInterfaceABI = `[
 						"type": "uint256"
 					}
 				],
-				"internalType": "struct IStaking.Candidate[]",
+				"internalType": "struct IStaking.Candidate",
 				"name": "",
-				"type": "tuple[]"
+				"type": "tuple"
 			}
 		],
 		"stateMutability": "view",
@@ -74,53 +69,47 @@ const _candidatesInterfaceABI = `[
 	}
 ]`
 
-var _candidatesMethod abi.Method
+var _candidateByNameMethod abi.Method
 
 func init() {
-	_interface, err := abi.JSON(strings.NewReader(_candidatesInterfaceABI))
+	_interface, err := abi.JSON(strings.NewReader(_candidateByNameInterfaceABI))
 	if err != nil {
 		panic(err)
 	}
 	var ok bool
-	_candidatesMethod, ok = _interface.Methods["candidates"]
+	_candidateByNameMethod, ok = _interface.Methods["candidateByName"]
 	if !ok {
 		panic("fail to load the method")
 	}
 }
 
-// CandidatesStateContext context for Candidates
-type CandidatesStateContext struct {
+// CandidateByNameStateContext context for CandidateByName
+type CandidateByNameStateContext struct {
 	*baseStateContext
 }
 
-func newCandidatesStateContext(data []byte) (*CandidatesStateContext, error) {
+func newCandidateByNameStateContext(data []byte) (*CandidateByNameStateContext, error) {
 	paramsMap := map[string]interface{}{}
 	ok := false
-	if err := _candidatesMethod.Inputs.UnpackIntoMap(paramsMap, data); err != nil {
+	if err := _candidateByNameMethod.Inputs.UnpackIntoMap(paramsMap, data); err != nil {
 		return nil, err
 	}
-	var offset, limit uint32
-	if offset, ok = paramsMap["offset"].(uint32); !ok {
-		return nil, errDecodeFailure
-	}
-	if limit, ok = paramsMap["limit"].(uint32); !ok {
+	var candName string
+	if candName, ok = paramsMap["candName"].(string); !ok {
 		return nil, errDecodeFailure
 	}
 
 	method := &iotexapi.ReadStakingDataMethod{
-		Method: iotexapi.ReadStakingDataMethod_CANDIDATES,
+		Method: iotexapi.ReadStakingDataMethod_CANDIDATE_BY_NAME,
 	}
 	methodBytes, err := proto.Marshal(method)
 	if err != nil {
 		return nil, err
 	}
 	arguments := &iotexapi.ReadStakingDataRequest{
-		Request: &iotexapi.ReadStakingDataRequest_Candidates_{
-			Candidates: &iotexapi.ReadStakingDataRequest_Candidates{
-				Pagination: &iotexapi.PaginationParam{
-					Offset: offset,
-					Limit:  limit,
-				},
+		Request: &iotexapi.ReadStakingDataRequest_CandidateByName_{
+			CandidateByName: &iotexapi.ReadStakingDataRequest_CandidateByName{
+				CandName: candName,
 			},
 		},
 	}
@@ -128,7 +117,7 @@ func newCandidatesStateContext(data []byte) (*CandidatesStateContext, error) {
 	if err != nil {
 		return nil, err
 	}
-	return &CandidatesStateContext{
+	return &CandidateByNameStateContext{
 		&baseStateContext{
 			&Parameters{
 				MethodName: methodBytes,
@@ -139,22 +128,18 @@ func newCandidatesStateContext(data []byte) (*CandidatesStateContext, error) {
 }
 
 // EncodeToEth encode proto to eth
-func (r *CandidatesStateContext) EncodeToEth(resp *iotexapi.ReadStateResponse) (string, error) {
-	var result iotextypes.CandidateListV2
+func (r *CandidateByNameStateContext) EncodeToEth(resp *iotexapi.ReadStateResponse) (string, error) {
+	var result iotextypes.CandidateV2
 	if err := proto.Unmarshal(resp.Data, &result); err != nil {
 		return "", err
 	}
 
-	args := make([]CandidateEth, len(result.Candidates))
-	for i, candidate := range result.Candidates {
-		cand, err := encodeCandidateToEth(candidate)
-		if err != nil {
-			return "", err
-		}
-		args[i] = *cand
+	cand, err := encodeCandidateToEth(&result)
+	if err != nil {
+		return "", err
 	}
 
-	data, err := _candidatesMethod.Outputs.Pack(args)
+	data, err := _candidateByNameMethod.Outputs.Pack(cand)
 	if err != nil {
 		return "", nil
 	}
